+++ wrappers/mvc-6/demos/Kendo.Mvc.Examples/Models/SamplesEntities/GanttResourceAssignment.cs
@@ -5,9 +5,9 @@ namespace Kendo.Mvc.Examples.Models
 {
     public partial class GanttResourceAssignment
     {
-        public long ID { get; set; }
-        public long ResourceID { get; set; }
-        public long TaskID { get; set; }
+        public int ID { get; set; }
+        public int ResourceID { get; set; }
+        public int TaskID { get; set; }
         public double Units { get; set; }
     }
 }
