+++ wrappers/mvc-6/demos/Kendo.Mvc.Examples/Models/SamplesEntities/GanttTask.cs
@@ -5,11 +5,11 @@ namespace Kendo.Mvc.Examples.Models
 {
     public partial class GanttTask
     {
-        public long ID { get; set; }
+        public int ID { get; set; }
         public string End { get; set; }
         public string Expanded { get; set; }
-        public long OrderID { get; set; }
-        public long? ParentID { get; set; }
+        public int OrderID { get; set; }
+        public int? ParentID { get; set; }
         public double PercentComplete { get; set; }
         public string Start { get; set; }
         public string Summary { get; set; }
